Se agrego soporte a numeros mas grandes

+++ src/models/Coins.java
@@ -1,4 +1,5 @@
 package models;
+import java.math.BigDecimal;
 
-public record Coins(String base_code, String target_code, double conversion_result) {
+public record Coins(String base_code, String target_code, BigDecimal conversion_result) {
 }
