+++ src/models/InformationConversion.java
@@ -1,18 +1,20 @@
 package models;
 
+import java.math.BigDecimal;
+
 public class InformationConversion {
     private String baseCode;
     private String targetCode;
-    private double amount;
-    private double conversionResult;
+    private BigDecimal amount;
+    private BigDecimal conversionResult;
 
-    public InformationConversion(String baseCode, String targetCode, double amount) {
+    public InformationConversion(String baseCode, String targetCode, BigDecimal amount) {
         this.baseCode = baseCode;
         this.targetCode = targetCode;
         this.amount = amount;
     }
 
-    public InformationConversion(Coins conversion, double amount) {
+    public InformationConversion(Coins conversion, BigDecimal amount) {
         this.baseCode = conversion.base_code();
         this.targetCode = conversion.target_code();
         this.conversionResult = conversion.conversion_result();
@@ -27,11 +29,11 @@ public class InformationConversion {
         return targetCode;
     }
 
-    public double getConversionResult() {
+    public BigDecimal getConversionResult() {
         return conversionResult;
     }
 
-    public double getAmount() {
+    public BigDecimal getAmount() {
         return amount;
     }
 }
